requirements.txt
alembic==1.13.1
asgiref==3.8.1
asyncpg==0.30.0
celery==5.4.0
fastapi[standard]==0.116.1
fastapi-mail==1.4.1
flower==2.0.1
httpx==0.28.1
isort==6.0.1
itsdangerous==2.2.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
PyJWT==2.8.0
redis==6.4.0
ruff==0.12.10
sqlmodel==0.0.24
greenlet==3.0.0
bcrypt==4.3.0
uvicorn==0.35.0
